add more tests for SC access/fence consistency
